Move query evaluation collection to corea
