+++ api/dbadapters/index.ts
@@ -5,8 +5,7 @@ import { BigQueryDbAdapter } from "df/api/dbadapters/bigquery";
 import { RedshiftDbAdapter } from "df/api/dbadapters/redshift";
 import { SnowflakeDbAdapter } from "df/api/dbadapters/snowflake";
 import { SQLDataWarehouseDBAdapter } from "df/api/dbadapters/sqldatawarehouse";
-import { ErrorWithCause } from "df/common/errors/errors";
-import { concatenateQueries } from "df/core/tasks";
+import { QueryOrAction } from "df/core/adapters";
 import { dataform } from "df/protos/ts";
 
 export type OnCancel = (handleCancel: () => void) => void;
@@ -71,75 +70,3 @@ register("postgres", RedshiftDbAdapter);
 register("redshift", RedshiftDbAdapter);
 register("snowflake", SnowflakeDbAdapter);
 register("sqldatawarehouse", SQLDataWarehouseDBAdapter);
-
-export type QueryOrAction = string | dataform.Table | dataform.Operation | dataform.Assertion;
-
-export function collectEvaluationQueries(
-  queryOrAction: QueryOrAction,
-  concatenate: boolean,
-  queryModifier: (mod: string) => string = (q: string) => q
-): dataform.ValidationQuery[] {
-  // TODO: The prefix method (via `queryModifier`) is a bit sketchy. For example after
-  // attaching the `explain` prefix, a table or operation could look like this:
-  // ```
-  // explain
-  // -- Delete the temporary table, if it exists (perhaps from a previous run).
-  // DROP TABLE IF EXISTS "df_integration_test"."load_from_s3_temp" CASCADE;
-  // ```
-  // which is invalid because the `explain` is interrupted by a comment.
-  const validationQueries = new Array<dataform.IValidationQuery>();
-  if (typeof queryOrAction === "string") {
-    validationQueries.push({ query: queryModifier(queryOrAction) });
-  } else {
-    try {
-      if (queryOrAction instanceof dataform.Table) {
-        if (queryOrAction.type === "incremental") {
-          const incrementalTableQueries = queryOrAction.incrementalPreOps.concat(
-            queryOrAction.incrementalQuery,
-            queryOrAction.incrementalPostOps
-          );
-          if (concatenate) {
-            validationQueries.push({
-              query: concatenateQueries(incrementalTableQueries, queryModifier),
-              incremental: true
-            });
-          } else {
-            incrementalTableQueries.forEach(q =>
-              validationQueries.push({ query: queryModifier(q), incremental: true })
-            );
-          }
-        }
-        const tableQueries = queryOrAction.preOps.concat(
-          queryOrAction.query,
-          queryOrAction.postOps
-        );
-        if (concatenate) {
-          validationQueries.push({
-            query: concatenateQueries(tableQueries, queryModifier)
-          });
-        } else {
-          tableQueries.forEach(q => validationQueries.push({ query: queryModifier(q) }));
-        }
-      } else if (queryOrAction instanceof dataform.Operation) {
-        if (concatenate) {
-          validationQueries.push({
-            query: concatenateQueries(queryOrAction.queries, queryModifier)
-          });
-        } else {
-          queryOrAction.queries.forEach(q => validationQueries.push({ query: queryModifier(q) }));
-        }
-      } else if (queryOrAction instanceof dataform.Assertion) {
-        validationQueries.push({ query: queryModifier(queryOrAction.query) });
-      } else {
-        throw new Error("Unrecognized evaluate type.");
-      }
-    } catch (e) {
-      throw new ErrorWithCause(`Error building tasks for evaluation. ${e.message}`, e);
-    }
-  }
-  return validationQueries
-    .map(validationQuery =>
-      dataform.ValidationQuery.create({ query: validationQuery.query.trim(), ...validationQuery })
-    )
-    .filter(validationQuery => !!validationQuery.query);
-}
